Finished factoring out the common LR(k) code. All that's left to do is some final cleanup and deletion of dead-code.
Note that it may also be possible to factor out the actual parse
function (which has nothing to do with the construction of the parse
table), but for now we shall leave it as is until it becomes clear that
we can in fact do this.

Finally, the LR(0) parser is actually incomplete, but this can wait
since the LR(0) module is not intended for practical use in any case.
(It is intended for conceptual and testing purposes only since it is
much simpler than the more complicated LR(k) parsers).